More device filtering logic to its own module
And switch from a long case statement to individual functions.
